Comm

+++ app/src/main/java/tramais/hnb/hhrfid/util/DisplayUtil.kt
@@ -38,9 +38,9 @@ object DisplayUtil {
     /**
      * 根据手机的分辨率从 dp 的单位 转成为 px(像素)
      */
-    fun dip2px(context: Context, dpValue: Float): Int {
+    fun dip2px(context: Context, dpValue: Float): Float {
         val scale: Float = context.resources.displayMetrics.density
-        return (dpValue * scale + 0.5f).toInt()
+        return (dpValue * scale + 0.5f)
     }
 
     /**
